go/consensus/tendermint: Move snapshot logic to its own file
